+++ Runtime/ServiceQuery.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace DGP.ServiceLocator
+{
+    internal class ServiceQuery
+    {
+        public ServiceSearchMode SearchMode;
+        public ServiceAddress Address;
+        public Action<ILocatableService> CallbackFn;
+    }
+}